move qwen36 moe vmm layer setup
